fix: improve multiline string handling
Multiline strings were being indented  like regular syntax this meant
that selecting the entire file or a large section and indenting it
would do crazy things to the multiline string content

I'm not sure of the best way to deal with multiline strings, but this
at least seems to stop arbitrary changes happening.

It's still not perfect, as it can't indent the closing quotes correctly.